requirements.txt
streamlit
deepface
numpy
pillow
opencv-python 